qubitapi/requirements.txt
aniso8601==8.0.0
bcrypt==3.1.7
boto3==1.12.42
botocore==1.15.42
cffi==1.14.0
click==7.1.1
docutils==0.15.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-RESTful==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
MarkupSafe==1.1.1
mysql-connector==2.2.9
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.9
Werkzeug==1.0.1
